Move some non-requesting file actions to a more appropriate spot
